+++ lenstronomy/LensModel/LineOfSight/single_plane_los_flexion.py
@@ -134,7 +134,8 @@ class SinglePlaneLOSFlexion(SinglePlane):
             the lens model classes
         :return: fermat potential in arcsec**2 as a list
         """
-        # Beware! Here the formula used for the computation of the fermat potential is different than the one presented in the paper "Weak lensing of strong lensing: beyond the tidal regime",
+        # Beware! Here the formula used for the computation of the fermat potential is different than the one presented in the paper 
+        # "Weak lensing of strong lensing: beyond the tidal regime",
         # but it is equivalent. This form has the advantage of being more compact.
 
         kwargs_main, kwargs_los_flexion = self.split_lens_los_flexion(kwargs_lens)
