+++ test/test_LensModel/test_LineOfSight/test_LOSFlexionModels/test_losflexion_minimal.py
@@ -5,9 +5,7 @@ import numpy.testing as npt
 import pytest
 import unittest
 
-from lenstronomy.LensModel.LineOfSightFlexion.LOSFlexionModels.losflexion_minimal import (
-    LOSFlexionMinimal,
-)
+from lenstronomy.LensModel.LineOfSight.LOSFlexionModels.losflexion_minimal import LOSFlexionMinimal
 
 
 class TestLOSFlexionMinimal(object):
